requirements.txt
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
Pillow==10.0.0
psycopg2-binary==2.9.9
django-filter==23.3
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.5.0
django-redis==5.4.0
redis==5.0.1
